+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 5.10.3 - Unreleased
 
 - Moving assets between folders now triggers a cache refresh ([#784](https://github.com/putyourlightson/craft-blitz/issues/784)).
+- Reverted the minimum plugin version requirement of 5.9.0 to help ease upgrading.
 
 ## 5.10.2 - 2025-04-09